requirements.txt
Flask~=3.1.0
pillow~=10.4.0
transformers~=4.38.2
langchain~=0.1.11
requests==2.31.0
torch==2.2.1